Update __init__.py
The {'class': 'level'} makes ignore two Supplies which hold a class : "bonus", so it returns just 8 rather than 10.  This alters the order and makes to report a wrong status for 'is_posible' in metal_storage,crystal_storage,deuterium_storage.

Deleting {'class':'level'} and reassigning index for Supplies class fix it.